Add 'call to action' and 'sign-up' buttons
This commit adds the call to action at the bottom of the page and
the sign-up button. What I need to do next is add a blue box around
these elements and then complete the footer. I have so far been
unsuccessful in creating the blue box which is proving to be a
challenge.