Correct blue button and quote
This commit amends certain things that were wrong with the project!
Firstly, it makes it so that the button will display in the center
no matter the zoom level of the UI. Secondly, it makes it so that
the name of the person who said the quote aligns with the quote
itself properly.