Re-correct blue-box and sign-up2
This commit makes it so that the box at the bottom scales correctly
no matter what zoom level. It also means that the elements within
scale appropriately.

If I were to continue the project, I would work out how to do this
with the header as those elements currently scale left and right,
but I think I am happy with the result as it is.